refactering states done
